Subject: [Implementation] Complete DeviceManager
Body:
Completes the DeviceManager implementation, providing all the functionality currently needed. Currently serving as the subject of observers. However, there is an outstanding task in the FileManager module regarding the implementation of TemporaryManager

Currently in the planning phase for Replacer, a critical core component with the highest priority. The design and development of Replacer are underway to ensure its effectiveness in meeting project requirements.

TODO:
The replacer. And complete the implementation of TemporaryManager in FileManager.

Note:
When adding new observer events, it is crucial to add them to the legal_labels within the Event module. Additionally, make sure to maintain consistency between the labels and the closure names used within the update function of the Interactor module.